(Sometimes) prevent device lockup upon object deletion.
Before, if a VJoyDevice() were del'd/GC'd, the device associated with it could not be modified any further until the Python instance were restarted.
This commit *sometimes* successfully releases it cleanly, but it seems there's something funky going on with _sdk.RelinquishVJD(rID) -- repeatedly acquiring and relinquishing a device causes it to eventually get locked out (either locked to a process without the process actually being able to access it, or ostensibly free but unable to be written to when acquired by a process). _sdk module seems to properly call the API, so this may be an issue with vJoy itself.